POST/DEV: Jekyll include post; StackEdit, code tweaks
Add post describing Jekyll {% include %}, along with img.html
as an example.

Revise styling for freestanding code blocks, to avoid the annoying
double-boxing effect. Achieved by narrowing the scope of the
'highlight' style to only <div>, not all tag types, so that the
<pre> doesn't also get it applied. Hopefully doesn't have annoying
knock-on effects.

Also took the chance to reduce the font size slighly in the code
blocks, to (hopefully) improve readability.

Rework the 'Written in StackEdit' notice to be an include, with
a dedicated style so that the font-size change actually is
rendered in the pages.