Clean up whitespace in rendered HTML on tags.md
Shouldn't affect the displayed content, but
makes the rendered HTML much cleaner on the tags page. Can
actually sort of see the structure of what's going on in the
render.

Definitely need to default to no-whitespace on the Liquid tags,
only allowing it in specific places to avoid run-on markup.